+++ src/blockchain.rs
@@ -169,7 +169,7 @@ pub trait Transactions {
 /// a fee strategy can be: fixed or range. When the fee strategy allows multiple possibilities, a
 /// [`FeePriority`] is used to determine what to apply.
 ///
-/// A fee strategy is included in an offer, so Alice and Bob can verify that transactions are valid
+/// A fee strategy is included in a deal, so Alice and Bob can verify that transactions are valid
 /// upon reception by the other participant.
 #[derive(Debug, Clone, Copy, Hash, Eq, PartialEq, Serialize, Deserialize)]
 pub enum FeeStrategy<T> {
@@ -383,8 +383,8 @@ impl FromStr for FeePriority {
 /// in fee strategies and an amount used in transactions. Implementing this trait allow to set and
 /// verify fees on transactions given a strategy and a priority.
 ///
-/// The fee to apply on transactions is carried in the [`Offer`](crate::negotiation::Offer) through
-/// a [`FeeStrategy`], in case the fee strategy allow multiple values a [`FeePriority`] is used to
+/// The fee to apply on transactions is carried in the [`Deal`](crate::trade::Deal) through a
+/// [`FeeStrategy`], in case the fee strategy allow multiple values a [`FeePriority`] is used to
 /// fix the amount.
 ///
 /// ```