+++ src/hash.rs
@@ -43,28 +43,3 @@ impl<'de> Visitor<'de> for HashString {
         }
     }
 }
-
-/// A visitor that deserializes a public offer
-pub(crate) struct OfferString;
-
-impl<'de> Visitor<'de> for OfferString {
-    type Value = String;
-
-    fn expecting(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
-        write!(
-            formatter,
-            "a string representing a public offer in Base58 value"
-        )
-    }
-
-    fn visit_str<E>(self, s: &str) -> Result<Self::Value, E>
-    where
-        E: de::Error,
-    {
-        if s.len() == 195 {
-            Ok(s.to_string())
-        } else {
-            Err(de::Error::invalid_length(s.len(), &self))
-        }
-    }
-}