+++ src/swap/btcxmr.rs
@@ -28,8 +28,8 @@ use crate::crypto::{
 #[cfg(feature = "experimental")]
 use crate::crypto::{EncSign, RecoverSecret, Sign};
 use crate::monero::Monero;
-use crate::negotiation;
 use crate::protocol;
+use crate::trade;
 use crate::{blockchain::Blockchain, crypto::dleq::DLEQProof};
 
 use monero::cryptonote::hash::Hash;
@@ -84,12 +84,12 @@ pub type Alice = protocol::Alice<bitcoin::Address, BitcoinSegwitV0, Monero>;
 /// Fully defined type for Bitcoin-Monero atomic swap Bob protocol role.
 pub type Bob = protocol::Bob<bitcoin::Address, BitcoinSegwitV0, Monero>;
 
-/// Fully defined type for Bitcoin-Monero atomic swap offer.
-pub type Offer = negotiation::Offer<bitcoin::Amount, monero::Amount, CSVTimelock, SatPerVByte>;
+/// Fully defined type for Bitcoin-Monero atomic swap trade.
+pub type DealParameters =
+    trade::DealParameters<bitcoin::Amount, monero::Amount, CSVTimelock, SatPerVByte>;
 
-/// Fully defined type for Bitcoin-Monero atomic swap public offer.
-pub type PublicOffer =
-    negotiation::PublicOffer<bitcoin::Amount, monero::Amount, CSVTimelock, SatPerVByte>;
+/// Fully defined type for Bitcoin-Monero atomic swap public trade.
+pub type Deal = trade::Deal<bitcoin::Amount, monero::Amount, CSVTimelock, SatPerVByte>;
 
 /// Fully defined type for Bitcoin-Monero atomic swap arbitrating parameters.
 pub type ArbitratingParameters =