+++ tests/protocol.rs
@@ -26,7 +26,7 @@ use farcaster_core::crypto::{
     ArbitratingKeyId, CommitmentEngine, GenerateKey, ProveCrossGroupDleq,
 };
 use farcaster_core::protocol::message::*;
-use farcaster_core::swap::btcxmr::{Alice, Bob, Parameters, PublicOffer};
+use farcaster_core::swap::btcxmr::{Alice, Bob, Deal, Parameters};
 use farcaster_core::swap::SwapId;
 use farcaster_core::transaction::*;
 
@@ -50,7 +50,7 @@ macro_rules! test_strict_ser {
     };
 }
 
-fn init() -> (Alice, Bob, PublicOffer) {
+fn init() -> (Alice, Bob, Deal) {
     let hex = "46435357415001004450e567b1106f429247bb680e5fe0c80200000080800000800800a08601000\
                00000000800c80000000000000004000a00000004000a0000000108001400000000000000022100\
                03b31a0a70343bb46f3db3768296ac5027f9873921b37f852860c690063ff9e4c90000000000000\
@@ -62,15 +62,14 @@ fn init() -> (Alice, Bob, PublicOffer) {
     let refund = Address::from_str("bc1qesgvtyx9y6lax0x34napc2m7t5zdq6s7xxwpvk").unwrap();
     let bob = Bob::new(Btc::new(), Xmr, refund, fee_politic);
 
-    let pub_offer: PublicOffer =
-        deserialize(&hex::decode(hex).unwrap()[..]).expect("Parsable public offer");
+    let deal: Deal = deserialize(&hex::decode(hex).unwrap()[..]).expect("Parsable deal");
 
-    (alice, bob, pub_offer)
+    (alice, bob, deal)
 }
 
 #[test]
 fn execute_offline_protocol() {
-    let (alice, bob, pub_offer) = init();
+    let (alice, bob, deal) = init();
 
     let commitment_engine = CommitmentEngine;
     let mut alice_key_manager = KeyManager::new(
@@ -97,14 +96,14 @@ fn execute_offline_protocol() {
     // Commit/Reveal round
     //
     let alice_params: Parameters = alice
-        .generate_parameters(&mut alice_key_manager, &pub_offer)
+        .generate_parameters(&mut alice_key_manager, &deal)
         .unwrap();
 
     let commit_alice_params = alice_params.commit_alice(swap_id, &commitment_engine);
     test_strict_ser!(commit_alice_params, CommitAliceParameters<KeccakCommitment>);
 
     let bob_params: Parameters = bob
-        .generate_parameters(&mut bob_key_manager, &pub_offer)
+        .generate_parameters(&mut bob_key_manager, &deal)
         .unwrap();
     let commit_bob_params = bob_params.commit_bob(swap_id, &commitment_engine);
     test_strict_ser!(commit_bob_params, CommitBobParameters<KeccakCommitment>);
@@ -154,7 +153,7 @@ fn execute_offline_protocol() {
             &alice_params,
             &bob_params,
             funding,
-            pub_offer.to_arbitrating_params(),
+            deal.to_arbitrating_params(),
         )
         .unwrap();
     let bob_cosign_cancel = bob
@@ -175,7 +174,7 @@ fn execute_offline_protocol() {
             &alice_params,
             &bob_params,
             &core,
-            pub_offer.to_arbitrating_params(),
+            deal.to_arbitrating_params(),
         )
         .unwrap();
     let cancel_sig = alice
@@ -184,7 +183,7 @@ fn execute_offline_protocol() {
             &alice_params,
             &bob_params,
             &core,
-            pub_offer.to_arbitrating_params(),
+            deal.to_arbitrating_params(),
         )
         .unwrap();
 
@@ -214,7 +213,7 @@ fn execute_offline_protocol() {
             &alice_params,
             &bob_params,
             &core,
-            pub_offer.to_arbitrating_params(),
+            deal.to_arbitrating_params(),
         )
         .unwrap();
     let signed_lock = bob
@@ -243,7 +242,7 @@ fn execute_offline_protocol() {
             &alice_params,
             &bob_params,
             &core,
-            pub_offer.to_arbitrating_params(),
+            deal.to_arbitrating_params(),
             &adaptor_buy,
         )
         .unwrap();
@@ -253,7 +252,7 @@ fn execute_offline_protocol() {
             &alice_params,
             &bob_params,
             &core,
-            pub_offer.to_arbitrating_params(),
+            deal.to_arbitrating_params(),
             &adaptor_buy,
         )
         .unwrap();
@@ -357,7 +356,7 @@ fn execute_offline_protocol() {
             &alice_params,
             &bob_params,
             &core,
-            pub_offer.to_arbitrating_params(),
+            deal.to_arbitrating_params(),
         )
         .unwrap();