+++ tests/trade.rs
@@ -20,8 +20,8 @@ use farcaster_core::blockchain::Blockchain;
 
 use farcaster_core::blockchain::{FeeStrategy, Network};
 use farcaster_core::consensus::{self, deserialize, serialize_hex};
-use farcaster_core::negotiation::{Offer, OfferFingerprint, PublicOffer};
 use farcaster_core::role::SwapRole;
+use farcaster_core::trade::{Deal, DealFingerprint, DealParameters};
 
 use bitcoin::Amount;
 use inet2_addr::InetSocketAddr;
@@ -30,37 +30,38 @@ use uuid::uuid;
 use std::str::FromStr;
 
 #[test]
-fn create_offer() {
+fn create_deal_parameters() {
     let hex = "4450e567b1106f429247bb680e5fe0c802000000808000008008000500000000000000080006000000000000000400070000000400080000000\
                10800090000000000000002";
-    let offer: Offer<bitcoin::Amount, monero::Amount, CSVTimelock, SatPerVByte> = Offer {
-        uuid: uuid!("67e55044-10b1-426f-9247-bb680e5fe0c8"),
-        network: Network::Testnet,
-        arbitrating_blockchain: Blockchain::Bitcoin,
-        accordant_blockchain: Blockchain::Monero,
-        arbitrating_amount: Amount::from_sat(5),
-        accordant_amount: monero::Amount::from_pico(6),
-        cancel_timelock: CSVTimelock::new(7),
-        punish_timelock: CSVTimelock::new(8),
-        fee_strategy: FeeStrategy::Fixed(SatPerVByte::from_sat(9)),
-        maker_role: SwapRole::Bob,
-    };
+    let deal: DealParameters<bitcoin::Amount, monero::Amount, CSVTimelock, SatPerVByte> =
+        DealParameters {
+            uuid: uuid!("67e55044-10b1-426f-9247-bb680e5fe0c8"),
+            network: Network::Testnet,
+            arbitrating_blockchain: Blockchain::Bitcoin,
+            accordant_blockchain: Blockchain::Monero,
+            arbitrating_amount: Amount::from_sat(5),
+            accordant_amount: monero::Amount::from_pico(6),
+            cancel_timelock: CSVTimelock::new(7),
+            punish_timelock: CSVTimelock::new(8),
+            fee_strategy: FeeStrategy::Fixed(SatPerVByte::from_sat(9)),
+            maker_role: SwapRole::Bob,
+        };
 
-    assert_eq!(hex, serialize_hex(&offer));
-    let strict_ser = strict_encoding::strict_serialize(&offer).unwrap();
+    assert_eq!(hex, serialize_hex(&deal));
+    let strict_ser = strict_encoding::strict_serialize(&deal).unwrap();
     assert_eq!(&hex::decode(hex).unwrap(), &strict_ser);
-    let res: Offer<bitcoin::Amount, monero::Amount, CSVTimelock, SatPerVByte> =
+    let res: DealParameters<bitcoin::Amount, monero::Amount, CSVTimelock, SatPerVByte> =
         strict_encoding::strict_deserialize(&strict_ser).unwrap();
-    assert_eq!(&offer, &res);
+    assert_eq!(&deal, &res);
 }
 
 #[test]
-fn get_offer_fingerprint() {
+fn get_deal_parameters_fingerprint() {
     let hex = "4450e567b1106f429247bb680e5fe0c802000000808000008008000500000000000000080006000\
                00000000000040007000000040008000000010800090000000000000002";
-    let res: Offer<bitcoin::Amount, monero::Amount, CSVTimelock, SatPerVByte> =
+    let res: DealParameters<bitcoin::Amount, monero::Amount, CSVTimelock, SatPerVByte> =
         strict_encoding::strict_deserialize(&hex::decode(hex).unwrap()).unwrap();
-    let id = OfferFingerprint::from_str(
+    let id = DealFingerprint::from_str(
         "f79b29ccb233b37cea3aa35b94c5ece25c58a8098afc18f046810a3c04591599",
     )
     .unwrap();
@@ -68,39 +69,40 @@ fn get_offer_fingerprint() {
     // other uuid
     let hex = "4351e567b1106f429247bb680e5fe0c802000000808000008008000500000000000000080006000\
                00000000000040007000000040008000000010800090000000000000002";
-    let res: Offer<bitcoin::Amount, monero::Amount, CSVTimelock, SatPerVByte> =
+    let res: DealParameters<bitcoin::Amount, monero::Amount, CSVTimelock, SatPerVByte> =
         strict_encoding::strict_deserialize(&hex::decode(hex).unwrap()).unwrap();
     // same fingerprint
     assert_eq!(id, res.fingerprint());
 }
 
 #[test]
-fn get_offer_uuid() {
+fn get_deal_parameters_uuid() {
     let hex = "4450e567b1106f429247bb680e5fe0c802000000808000008008000500000000000000080006000\
                00000000000040007000000040008000000010800090000000000000002";
-    let res: Offer<bitcoin::Amount, monero::Amount, CSVTimelock, SatPerVByte> =
+    let res: DealParameters<bitcoin::Amount, monero::Amount, CSVTimelock, SatPerVByte> =
         strict_encoding::strict_deserialize(&hex::decode(hex).unwrap()).unwrap();
     assert_eq!(uuid!("67e55044-10b1-426f-9247-bb680e5fe0c8"), res.uuid());
 }
 
 #[test]
-fn serialize_public_offer() {
+fn serialize_deal() {
     let hex = "46435357415001004450e567b1106f429247bb680e5fe0c80200000080800000800800a08601000\
                00000000800c80000000000000004000a00000004000a0000000108001400000000000000022100\
                03b31a0a70343bb46f3db3768296ac5027f9873921b37f852860c690063ff9e4c90000000000000\
                0000000000000000000000000000000000000000000000000000000260700";
-    let offer: Offer<bitcoin::Amount, monero::Amount, CSVTimelock, SatPerVByte> = Offer {
-        uuid: uuid!("67e55044-10b1-426f-9247-bb680e5fe0c8"),
-        network: Network::Testnet,
-        arbitrating_blockchain: Blockchain::Bitcoin,
-        accordant_blockchain: Blockchain::Monero,
-        arbitrating_amount: Amount::from_sat(100000),
-        accordant_amount: monero::Amount::from_pico(200),
-        cancel_timelock: CSVTimelock::new(10),
-        punish_timelock: CSVTimelock::new(10),
-        fee_strategy: FeeStrategy::Fixed(SatPerVByte::from_sat(20)),
-        maker_role: SwapRole::Bob,
-    };
+    let deal: DealParameters<bitcoin::Amount, monero::Amount, CSVTimelock, SatPerVByte> =
+        DealParameters {
+            uuid: uuid!("67e55044-10b1-426f-9247-bb680e5fe0c8"),
+            network: Network::Testnet,
+            arbitrating_blockchain: Blockchain::Bitcoin,
+            accordant_blockchain: Blockchain::Monero,
+            arbitrating_amount: Amount::from_sat(100000),
+            accordant_amount: monero::Amount::from_pico(200),
+            cancel_timelock: CSVTimelock::new(10),
+            punish_timelock: CSVTimelock::new(10),
+            fee_strategy: FeeStrategy::Fixed(SatPerVByte::from_sat(20)),
+            maker_role: SwapRole::Bob,
+        };
     let ip = FromStr::from_str("0.0.0.0").unwrap();
     let port = FromStr::from_str("9735").unwrap();
 
@@ -112,25 +114,25 @@ fn serialize_public_offer() {
     .inner;
     let node_id = secp256k1::PublicKey::from_secret_key(&secp, &sk);
     let peer_address = InetSocketAddr::socket(ip, port);
-    let public_offer = offer.to_public_v1(node_id, peer_address);
+    let deal = deal.to_v1(node_id, peer_address);
 
-    assert_eq!(hex, serialize_hex(&public_offer));
-    let strict_ser = strict_encoding::strict_serialize(&public_offer).unwrap();
+    assert_eq!(hex, serialize_hex(&deal));
+    let strict_ser = strict_encoding::strict_serialize(&deal).unwrap();
     assert_eq!(&hex::decode(hex).unwrap(), &strict_ser);
-    let res: PublicOffer<bitcoin::Amount, monero::Amount, CSVTimelock, SatPerVByte> =
+    let res: Deal<bitcoin::Amount, monero::Amount, CSVTimelock, SatPerVByte> =
         strict_encoding::strict_deserialize(&strict_ser).unwrap();
-    assert_eq!(&public_offer, &res);
+    assert_eq!(&deal, &res);
 }
 
 #[test]
-fn get_public_offer_fingerprint() {
+fn get_deal_fingerprint() {
     let hex = "46435357415001004450e567b1106f429247bb680e5fe0c80200000080800000800800a08601000\
                00000000800c80000000000000004000a00000004000a0000000108001400000000000000022100\
                03b31a0a70343bb46f3db3768296ac5027f9873921b37f852860c690063ff9e4c90000000000000\
                0000000000000000000000000000000000000000000000000000000260700";
-    let res: PublicOffer<bitcoin::Amount, monero::Amount, CSVTimelock, SatPerVByte> =
+    let res: Deal<bitcoin::Amount, monero::Amount, CSVTimelock, SatPerVByte> =
         strict_encoding::strict_deserialize(&hex::decode(hex).unwrap()).unwrap();
-    let id = OfferFingerprint::from_str(
+    let id = DealFingerprint::from_str(
         "3a466a0a0cff7bf800808653460076549621d07db78e697b9dfaebaba0ab8b33",
     )
     .unwrap();
@@ -140,48 +142,48 @@ fn get_public_offer_fingerprint() {
                00000000800c80000000000000004000a00000004000a0000000108001400000000000000022100\
                03b31a0a70343bb46f3db3768296ac5027f9873921b37f852860c690063ff9e4c90000000000000\
                0000000000000000000000000000000000000000000000000000000260700";
-    let res: PublicOffer<bitcoin::Amount, monero::Amount, CSVTimelock, SatPerVByte> =
+    let res: Deal<bitcoin::Amount, monero::Amount, CSVTimelock, SatPerVByte> =
         strict_encoding::strict_deserialize(&hex::decode(hex).unwrap()).unwrap();
     // same fingerprint
     assert_eq!(id, res.fingerprint());
 }
 
 #[test]
-fn get_public_offer_uuid() {
+fn get_deal_uuid() {
     let hex = "46435357415001004450e567b1106f429247bb680e5fe0c80200000080800000800800a08601000\
                00000000800c80000000000000004000a00000004000a0000000108001400000000000000022100\
                03b31a0a70343bb46f3db3768296ac5027f9873921b37f852860c690063ff9e4c90000000000000\
                0000000000000000000000000000000000000000000000000000000260700";
-    let res: PublicOffer<bitcoin::Amount, monero::Amount, CSVTimelock, SatPerVByte> =
+    let res: Deal<bitcoin::Amount, monero::Amount, CSVTimelock, SatPerVByte> =
         strict_encoding::strict_deserialize(&hex::decode(hex).unwrap()).unwrap();
     assert_eq!(uuid!("67e55044-10b1-426f-9247-bb680e5fe0c8"), res.uuid());
 }
 
 #[test]
-fn check_public_offer_magic_bytes() {
+fn check_deal_magic_bytes() {
     let valid = "46435357415001004450e567b1106f429247bb680e5fe0c80200000080800000800800a08601000\
                  00000000800c80000000000000004000a00000004000a0000000108001400000000000000022100\
                  03b31a0a70343bb46f3db3768296ac5027f9873921b37f852860c690063ff9e4c90000000000000\
                  0000000000000000000000000000000000000000000000000000000260700";
-    let pub_offer: Result<
-        PublicOffer<bitcoin::Amount, monero::Amount, CSVTimelock, SatPerVByte>,
+    let deal: Result<
+        Deal<bitcoin::Amount, monero::Amount, CSVTimelock, SatPerVByte>,
         consensus::Error,
     > = deserialize(&hex::decode(valid).unwrap()[..]);
-    assert!(pub_offer.is_ok());
+    assert!(deal.is_ok());
 
     let invalid = "47435357415001004450e567b1106f429247bb680e5fe0c80200000080800000800800a08601000\
                  00000000800c80000000000000004000a00000004000a0000000108001400000000000000022100\
                  03b31a0a70343bb46f3db3768296ac5027f9873921b37f852860c690063ff9e4c90000000000000\
                  0000000000000000000000000000000000000000000000000000000260700";
-    let pub_offer: Result<
-        PublicOffer<bitcoin::Amount, monero::Amount, CSVTimelock, SatPerVByte>,
+    let deal: Result<
+        Deal<bitcoin::Amount, monero::Amount, CSVTimelock, SatPerVByte>,
         consensus::Error,
     > = deserialize(&hex::decode(invalid).unwrap()[..]);
-    assert!(pub_offer.is_err());
+    assert!(deal.is_err());
 }
 
 #[test]
-fn parse_public_offer() {
+fn parse_deal() {
     for hex in [
         "46435357415001004450e567b1106f429247bb680e5fe0c80200000080800000800800a08601000\
          00000000800c80000000000000004000a00000004000a0000000108001400000000000000022100\
@@ -191,7 +193,7 @@ fn parse_public_offer() {
     .iter_mut()
     {
         let bytes = hex::decode(hex).expect("hex");
-        let res: Result<PublicOffer<bitcoin::Amount, monero::Amount, CSVTimelock, SatPerVByte>, _> =
+        let res: Result<Deal<bitcoin::Amount, monero::Amount, CSVTimelock, SatPerVByte>, _> =
             strict_encoding::strict_deserialize(&bytes);
         assert!(res.is_ok());
     }